Changed std::variant<int, float to std::variant<int, double> in line 41:

+++ header-source/headers/model-headers/decision-tree-header/tree-header.hpp
@@ -38,7 +38,7 @@ struct SplitYieldParameters {
 struct GeneratorVariables {
     std::string                 split_type;
     int                         split_index;
-    std::variant<int, float>    best_temporary_numeric_threshold;
+    std::variant<int, double>   best_temporary_numeric_threshold;
     int                         best_temporary_categorical_threshold;
     arma::mat&                  left_x_subset;
     arma::vec&                  left_y_subset;
